Refactored request handling
